Create new port php5-xslcache
The XSL Cache extension is a modification of PHP's standard XSL extension that
caches the parsed XSL stylesheet representation between sessions for 2.5x boost
in performance for sites that repeatedly apply the same transform. Although
there is still some further work that could be done on the extension, this code
is already proving beneficial in production use for a few applications on the
New York Times' website.


git-svn-id: https://svn.macports.org/repository/macports/trunk@62324 d073be05-634f-4543-b044-5fe20cf6d1d6