Create new port php5-big_int
A big_int library provides a set of functions for calculations with
arbitrary length integers and bitsets. Functions from this package are
useful for number theory applications. For example, in two-keys
cryptography. The package has many bitset functions, which allow to work
with arbitrary length bitsets. This package is much faster than bundled
into PHP BCMath and consists almost all functions, which are implemented
in PHP GMP extension, but it needn't any external libraries.


git-svn-id: https://svn.macports.org/repository/macports/trunk@62597 d073be05-634f-4543-b044-5fe20cf6d1d6