py-game: use sha256 checksum
The actual purpose of this change is to force portindex to rescan this
file. The py27-game subport was not being included in the index
because there was a separate py27-game port; now that that's been
removed (r84128), we need to reindex this port to add the subport to
the index.


git-svn-id: https://svn.macports.org/repository/macports/trunk@84169 d073be05-634f-4543-b044-5fe20cf6d1d6