Don't rely on mtime of statefile and Portfile
Traditionally, MacPorts used mtimes of both the statefile and the
Portfile to detect whether a Portfile was updated between builds and a
previously started build should not be continued, but restarted.

However, since we do sync modification times from the repository to the
users there exists a situation where this check is not correct (see
#29223 for an explanation).

This change drops checking the mtime in favor of recording a sha256 sum
of the Portfile when starting a new build as first line of the
statefile. If the checksum doesn't match, the build is restarted.

Since a check for changed variants did rely on the statefile being empty
for fresh builds, I have re-written this test to consider the statefile
"empty" if no lines match "variant: .*".

This also updates man 1 port.

All flags affecting the check (-f, -o) should continue to work.

The change of the statefile format will cause all builds started before
installing this code to be restarted.

Closes #29223.


git-svn-id: https://svn.macports.org/repository/macports/trunk@96136 d073be05-634f-4543-b044-5fe20cf6d1d6