base: Keep supporting old statefile layout, followup to r96136, #29223
Dropping support for old statefiles and just re-starting the build had
some undesired effects, e.g., when installing from binary archives
generated with the older statefile format or when trying to make a
pkg/mpkg from a previously installed package with the old statefile
format.

This change re-introduces support for reading old statefiles to fix
these problems. It also adds a version number to the statefile and warns
for unknown statefile versions.


git-svn-id: https://svn.macports.org/repository/macports/trunk@96399 d073be05-634f-4543-b044-5fe20cf6d1d6