py2{5,6,7}-lint: new upstream 0.25.2 release.
Delete the patch as it appears it is unnecessary.  From a quick
review, it appears that ${prefix}/bin/epylint-2.X imports epylint from
the site-packages directory which then finds the 'lint.py' to run
using

    cmd = [sys.executable, lintPath, '-f', 'parseable', '-r', 'n',
           '--disable=C,R,I', childPath]
    process = Popen(cmd, stdout=PIPE, stderr=PIPE, cwd=parentPath)

which will get the correct 'lint.py', which will be in each different
Python version's site-packages directory.


git-svn-id: https://svn.macports.org/repository/macports/trunk@96602 d073be05-634f-4543-b044-5fe20cf6d1d6