Automatically add dependency on MacPorts-provided compilers, if chosen as configure.compiler. Closes #32542.
This also adds a generic mechanism to register callbacks to be run after
sourcing the Portfile and evaluating the variants. This callback
mechanism can probably be used to simplify other parts of port1.0, e.g.
automatically adding a dependency on autoconf when use_autoreconf is set
to yes, automatically adding a dependency on bsdmake when build.type is
bsd, etc.

For ports that blacklist all Xcode-provided compilers using
compiler.blacklist (so a fallback from MacPorts is used) the dependency
information recoded in the PortIndex will be missing these automatic
dependencies, because the PortIndex is usually generated on a server
without Xcode.


git-svn-id: https://svn.macports.org/repository/macports/trunk@102932 d073be05-634f-4543-b044-5fe20cf6d1d6