dosbox:
Always autoreconf

I just hit a weird failure after starting a fresh install on all my machines
last night where I was stuck in a loop in build.  It looked like it detected
changes that required re-running autoconf, so it did that and failed because
of version-mismatches.  It then kept trying that forever.  Safest and easiest
change is to just autoreconf.


git-svn-id: https://svn.macports.org/repository/macports/trunk@107473 d073be05-634f-4543-b044-5fe20cf6d1d6