Change the CMAKE_BUILD_TYPE to Debug instead of debugFull for the +debug variant in the CMake PortGroup. (#40648)
"debugFull" is not a standard build type in CMake so there are no default compiler flags set for 
it, which would turn on the inclusion of debug symbols. Instead, CMake provides the build type 
"Debug" for this purpose. Only KDE provides in its CMake build infrastructure support for a build 
type called "DebugFull", which turns off optimization, as KDE leaves it on for the "Debug" build 
type. Confer to <​http://techbase.kde.org/Development/CMake/Addons_for_KDE> for further 
explanations.


git-svn-id: https://svn.macports.org/repository/macports/trunk@112837 d073be05-634f-4543-b044-5fe20cf6d1d6