port.tcl: switch progress bars to ascii-only, capsule in a subnamespace
The Unicode block drawing characters produce unsatisfactory results with the
default configurations of Terminal.app. This change reverts to ASCII-only
characters for the progress bars (even just using a single full block looks
ugly in some of the fonts in the default configurations).

Unfortunately it seems we'll have to wait until Unicode fonts provide better
support for the block drawing characters.

This change also prevents the progress bars from overflowing the width of the
terminal window and provides a namespace for progress functions to avoid
cluttering the global namespace with the required helper variables.


git-svn-id: https://svn.macports.org/repository/macports/trunk@117186 d073be05-634f-4543-b044-5fe20cf6d1d6