configure.ac: Reset $PATH to the default before configuring
... and remove some references to /usr/local that we really don't want. Also
make sure all binaries that are in /sbin or /usr/sbin have an appropriate PATH
hint in AC_CHECK_PROG, just in case somebody doesn't have the sbin directories
in his $PATH (which should never happen now that configure resets the PATH, but
better be safe than sorry).

This avoids MacPorts thinking this was a SystemStarter-system because it can't
find /sbin/launchd during configure.


git-svn-id: https://svn.macports.org/repository/macports/trunk@117518 d073be05-634f-4543-b044-5fe20cf6d1d6