libtool: Avoid unnecessary regeneration during build
Our patches cause the build process to regenerate several files that
don't really have to be regenerated. So the build dependencies on
autoconf and automake are not necessary, strictly speaking.

The build also runs the configure script a second time, during the build
phase. This second configure run detects MacPorts' grep and gsed if they
are available and subsequently pollutes glibtoolize (#45509).

Reworking the stdlib patch and updating modification dates on select
files solves these problems and lets us drop the build dependencies on
autoconf and automake.


git-svn-id: https://svn.macports.org/repository/macports/trunk@127198 d073be05-634f-4543-b044-5fe20cf6d1d6