machista: add missing include, re-generate with newer swig
This adds a warning about a missing initializer when compiling machista_wrap.c
with clang. This warning is a false positive. While there is no explicit
initializer for the last member of the struct (a Tcl hashtable), the C99
standard guarantees unitialized static structs are set to 0 (and the value
doesn't really matter anyway, because it will be properly initialized later
on). Putting a {0} there does not really help either, because clang doesn't
recognize this standard way to say "initialize the rest with zeroes".

=> I'm ignoring this warning.


git-svn-id: https://svn.macports.org/repository/macports/trunk@136929 d073be05-634f-4543-b044-5fe20cf6d1d6