base: rename pextlib flock to adv-flock to avoid conflict with TclX, closes #46733
This has come up a few times and is somehow related to the order in which the
packages are loaded, but I haven't been able to pinpoint why exactly the order
varies.

Let's avoid wasting any time on the dirty details and just rename our own
implementation. The new name is "adv-flock" for "advisory file lock".


git-svn-id: https://svn.macports.org/repository/macports/trunk@138165 d073be05-634f-4543-b044-5fe20cf6d1d6