archivefetch: no fallback to default archive_sites
If a ports tree provides a modified version of a port, falling back to the
default ports tree might fetch a binary archive that was built from a different
Portfile. As this binary archive would not have the expected contents and might
be incompatible, only the archive sites defined for this ports tree should be
checked.

Using any globally defined archive site from archive_sites.conf is fine, as we
expect the user to be aware that this has to match the ports trees as defined
in sources.conf.


git-svn-id: https://svn.macports.org/repository/macports/trunk@147456 d073be05-634f-4543-b044-5fe20cf6d1d6