Mass update of python ports for master_sites pypi
The pypi hosting site changed the naming scheme for new uploads. While old URLs
continue to work, this possibly breaks updates as master_sites would need to be
changed. This patch updates all previously hardcoded references to
pypi.python.org to the pypi: mirror sites, which was already updated to check
at both the old and new location. See #51391.

Overriding maintainers as this is a mass update of ports.


git-svn-id: https://svn.macports.org/repository/macports/trunk@148976 d073be05-634f-4543-b044-5fe20cf6d1d6