libbpg: update to 0.9.7
As of 0.9.6, x265 is used by default instead of jctvc. However, also as
of 0.9.6, a bundled copy of x265 is used instead of a system copy. That
alone would be undesirable from MacPorts' point of view, but the bundled
copy also fails to build, because it tries to enable "numa", which is
for Linux only. Therefore libbpg 0.9.7 in MacPorts continues to use
jctvc, not x265. I've asked the author of libbpg to give us an option to
use a system version of x265.


git-svn-id: https://svn.macports.org/repository/macports/trunk@153158 d073be05-634f-4543-b044-5fe20cf6d1d6